pyproject.toml
[tool.poetry]
name = "blog-coeur"
version = "0.0.12"
description = "Coeur - static site management"
authors = ["David Silva <srdavidsilva@gmail.com>"]
readme = "README.md"
repository = "https://github.com/sr2ds/coeur"
packages = [
    { include = "coeur" }
]

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.12.3"}
toml = "^0.10.2"
sqlalchemy = "^2.0.30"
jinja2 = "^3.1.4"
psutil = "^5.9.8"
mistune = "^3.0.2"
pyyaml = "^6.0.1"
watchdog = "^4.0.1"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
python-slugify = "^8.0.4"
minify-html = "^0.15.0"
instagrapi = "^2.1.2"
pillow = "^11.0.0"

[tool.poetry.scripts]
blog-coeur = "coeur.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"